Groundwork for evals
ikyk stands for "I know you know." It creates a bit table with only the
bits that both passed players know. In order to be able to create a
floating, temporary bit table, I modified the table init slightly.

In addition, there's now a space in Events to track which cards were
touched by a clue. This information could be derived rather than stored,
assuming full replay capability, but as long as we don't have that this
should make a lot of things easier.