First pass at test scripting
There's an example script, an example test deck, some modifications to
Classes to allow for testing (and also some more debugging stuff) and a
new testing-universe version of Hanabi. A script can now be executed on
the command line by importing HanabiTest and then calling
tst_play_hanabi, passing in the number associated with the test deck and
script you want to use.

The main Hanabi script was also modified because part of this change
involved pulling out a bunch of the HanabiGame init and calling it at
the Hanabi level. That enables the test script to swap in a test deck at
the right time, so that the card IDs all match up and etc.

Limitations:
-cluing is not yet implemented
-you still have to enter the player order yourself