some progress (see descrip)
I need to go become a person for a while, but here's the deal:
Yes, the format of the blank cards is kinda dumb, but they'll work if we
fix a few things:
The final_countdown needs to be triggered by drawing a 'blank card' ...
card.color = 'D'  is sufficient for this.
Need to prevent playing/discarding/cluing blank cards.  Can do this
through bumping -100000 or through going back and altering the way the
choice list is generated.